requirements/common.txt
anyio==3.4.0
asgiref==3.4.1
click==8.0.3
fastapi==0.71.0
h11==0.12.0
idna==3.3
peewee==3.14.8
pydantic==1.9.0
PyMySQL==1.0.2
sniffio==1.2.0
starlette==0.17.1
typing-extensions==4.0.1
uvicorn==0.16.0
python-jose[cryptography]
passlib[bcrypt]
fastapi_jwt_auth
python-multipart

celery==5.2.3
